--- Common/src/main/java/dk/sdu/sesem4/common/data/events/Event.java
@@ -1,4 +1,0 @@
-package dk.sdu.sesem4.common.data.events;
-
-public abstract class Event {
-}
