+++ Common/src/main/java/dk/sdu/sesem4/common/event/Event.java
@@ -1,0 +1,4 @@
+package dk.sdu.sesem4.common.event;
+
+public abstract class Event {
+}
