+++ Common/src/main/java/dk/sdu/sesem4/common/event/EventListener.java
@@ -1,4 +1,4 @@
-package dk.sdu.sesem4.common.data.events;
+package dk.sdu.sesem4.common.event;
 
 public interface EventListener {
     void processNotification(EventType eventType, Event data);
